SkinAI_frontend-main/SkinAI_frontend-main/requirements.txt
streamlit
requests
Pillow
numpy